refactor: move endpoints to the corrected positions
